# swagger.yaml
openapi: 3.0.3
info:
  title: BitCommerce OAuth API
  description: |
    # BitCommerce Merchant API with OAuth 2.0 Authorization Code Flow
    
    This API provides secure access to merchant data through OAuth 2.0 Authorization Code Flow.
    
    ## OAuth Flow Overview
    
    1. **App Registration** - Register your application to get client credentials
    2. **Authorization** - Redirect merchant to authorization page
    3. **Token Exchange** - Exchange authorization code for access tokens
    4. **API Access** - Use access token to access merchant data
    5. **Token Refresh** - Refresh expired access tokens
    
    ## Quick Start
    
    ```bash
    # 1. Register your application
    curl -X POST http://localhost:3000/oauth/apps \
      -H "Content-Type: application/json" \
      -d '{
        "name": "My Integration App",
        "redirect_uri": "https://myapp.com/callback",
        "scopes": ["read_products", "read_orders"]
      }'
    
    # 2. Redirect merchant to authorization URL
    # http://localhost:3000/oauth/authorize?client_id=YOUR_CLIENT_ID&redirect_uri=https://myapp.com/callback&scope=read_products&response_type=code
    
    # 3. Exchange authorization code for tokens
    curl -X POST http://localhost:3000/oauth/token \
      -H "Content-Type: application/json" \
      -d '{
        "grant_type": "authorization_code",
        "client_id": "YOUR_CLIENT_ID",
        "client_secret": "YOUR_CLIENT_SECRET",
        "code": "AUTH_CODE_FROM_REDIRECT",
...
    
...
servers:
...
paths:
...
  /api/v1/merchants/{merchant_id}/profile:
    get:
      tags:
        - Merchants
      summary: Get merchant profile
      description: |
        Get merchant profile information.
        
        **Required Scope:** `read_profile`
      parameters:
        - name: merchant_id
          in: path
          required: true
          schema:
            type: string
          description: Merchant ID
          example: "merchant_123"
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Merchant profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MerchantProfile'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OAuthError'
        '403':
          description: Forbidden (insufficient scope or merchant access)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OAuthError'
        '404':
          description: Merchant not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OAuthError'
      examples:
        curl:
          summary: Get merchant profile
          value: |
            curl -H "Authorization: Bearer YOUR_ACCESS_TOKEN" \
              http://localhost:3000/api/v1/merchants/merchant_123/profile
  /api/v1/merchants:
    get:
      tags:
        - Merchants
      summary: List all merchants
      security:
        - BearerAuth: []
      responses:
        '200':
          description: A list of merchants
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Merchant'
  /api/v1/merchants/{merchant_id}/products:
    get:
      tags:
        - Merchants
      summary: List products for a merchant
      security:
        - BearerAuth: []
      parameters:
        - name: merchant_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: A list of products
          content:
            application/json:
              schema:
                type: object
                properties:
                  merchant:
                    $ref: '#/components/schemas/Merchant'
                  products:
                    type: array
                    items:
                      $ref: '#/components/schemas/Product'
                  total:
                    type: integer
  /api/v1/merchants/{merchant_id}/orders:
    get:
      tags:
        - Merchants
      summary: List orders for a merchant
      security:
        - BearerAuth: []
      parameters:
        - name: merchant_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: A list of orders
          content:
            application/json:
              schema:
                type: object
                properties:
                  merchant:
                    $ref: '#/components/schemas/Merchant'
                  orders:
                    type: array
                    items:
                      $ref: '#/components/schemas/Order'
                  total:
                    type: integer
  /api/v1/merchants/{merchant_id}/products/{product_id}:
    get:
      tags:
        - Merchants
      summary: Get specific product
      description: |
        Get a specific product by ID.
        
        **Required Scope:** `read_products`
      parameters:
        - name: merchant_id
          in: path
          required: true
          schema:
            type: string
          description: Merchant ID
          example: "merchant_123"
        - name: product_id
          in: path
          required: true
          schema:
            type: string
          description: Product ID
          example: "product_1"
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Product details
          content:
            application/json:
              schema:
                type: object
                properties:
                  merchant:
                    $ref: '#/components/schemas/MerchantSummary'
                  product:
                    $ref: '#/components/schemas/Product'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OAuthError'
        '403':
          description: Forbidden (insufficient scope or merchant access)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OAuthError'
        '404':
          description: Merchant or product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OAuthError'
      examples:
        curl:
          summary: Get specific product
          value: |
            curl -H "Authorization: Bearer YOUR_ACCESS_TOKEN" \
              http://localhost:3000/api/v1/merchants/merchant_123/products/product_1
  /api/v1/merchants/{merchant_id}/orders/{order_id}:
    get:
      tags:
        - Merchants
      summary: Get specific order
      description: |
        Get a specific order by ID.
        
        **Required Scope:** `read_orders`
      parameters:
        - name: merchant_id
          in: path
          required: true
          schema:
            type: string
          description: Merchant ID
          example: "merchant_123"
        - name: order_id
          in: path
          required: true
          schema:
            type: string
          description: Order ID
          example: "order_1"
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Order details
          content:
            application/json:
              schema:
                type: object
                properties:
                  merchant:
                    $ref: '#/components/schemas/MerchantSummary'
                  order:
                    $ref: '#/components/schemas/Order'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OAuthError'
        '403':
          description: Forbidden (insufficient scope or merchant access)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OAuthError'
        '404':
          description: Merchant or order not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OAuthError'
      examples:
        curl:
          summary: Get specific order
          value: |
            curl -H "Authorization: Bearer YOUR_ACCESS_TOKEN" \
              http://localhost:3000/api/v1/merchants/merchant_123/orders/order_1
  /api/v1/merchants/{merchant_id}/summary:
    get:
      tags:
        - Merchants
      summary: Get merchant summary
      description: |
        Get a summary of merchant data including product and order statistics.
        
        **Required Scope:** `read_products` OR `read_orders`
      parameters:
        - name: merchant_id
          in: path
          required: true
          schema:
            type: string
          description: Merchant ID
          example: "merchant_123"
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Merchant summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MerchantSummary'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OAuthError'
        '403':
          description: Forbidden (insufficient scope or merchant access)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OAuthError'
        '404':
          description: Merchant not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OAuthError'
      examples:
        curl:
          summary: Get merchant summary
          value: |
            curl -H "Authorization: Bearer YOUR_ACCESS_TOKEN" \
              http://localhost:3000/api/v1/merchants/merchant_123/summary
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Include your access token in the Authorization header.
        
        Example: `Authorization: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...`
  schemas:
...
    OAuthError:
      type: object
      properties:
        error:
          type: string
          enum: [invalid_request, invalid_client, invalid_grant, access_denied, invalid_token, insufficient_scope, not_found, server_error]
          example: "invalid_request"
        error_description:
          type: string
          example: "Missing required parameter: client_id"
    MerchantProfile:
      type: object
      properties:
        id:
          type: string
          example: "merchant_123"
        name:
          type: string
          example: "Sample Store"
        email:
          type: string
          format: email
          example: "merchant@samplestore.com"
        status:
          type: string
          enum: [active, inactive, suspended]
          example: "active"
        created_at:
          type: string
          format: date-time
          example: "2024-01-01T00:00:00.000Z"
        updated_at:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00.000Z"
    MerchantSummary:
      type: object
      properties:
        id:
          type: string
          example: "merchant_123"
        name:
          type: string
          example: "Sample Store"
        status:
          type: string
          enum: [active, inactive, suspended]
          example: "active"
    Product:
      type: object
      properties:
        id:
          type: string
          example: "product_1"
        merchant_id:
          type: string
          example: "merchant_123"
        name:
          type: string
          example: "Sample Product 1"
        price:
          type: number
          format: float
          example: 29.99
        description:
          type: string
          example: "A sample product for testing"
        status:
          type: string
          enum: [active, inactive, draft]
          example: "active"
        created_at:
          type: string
          format: date-time
          example: "2024-01-01T00:00:00.000Z"
    Order:
      type: object
      properties:
        id:
          type: string
          example: "order_1"
        merchant_id:
          type: string
          example: "merchant_123"
        customer_email:
          type: string
          format: email
          example: "customer@example.com"
        total:
          type: number
          format: float
          example: 79.98
        status:
          type: string
          enum: [pending, processing, completed, cancelled]
          example: "completed"
        items:
          type: array
          items:
            type: object
            properties:
              product_id:
                type: string
                example: "product_1"
              quantity:
                type: integer
                example: 1
              price:
                type: number
                format: float
                example: 29.99
        created_at:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00.000Z"
    Merchant:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        status:
          type: string
        created_at:
          type: string
          format: date-time

# Complete OAuth Flow Example
...